+++ weldx/asdf/utils.py
@@ -275,11 +275,70 @@ def create_asdf_dataclass(
     return asdf_file_path, python_file_path
 
 
+def _write_buffer(tree: dict, asdffile_kwargs: dict = None, write_kwargs: dict = None):
+    """Write ASDF file into buffer.
+
+    Parameters
+    ----------
+    tree:
+        Tree object to serialize.
+    asdffile_kwargs
+        Additional keywords to pass to asdf.AsdfFile()
+    write_kwargs
+        Additional keywords to pass to asdf.AsdfFile.write_to()
+        Weldx-Extensions are always set.
+
+    Returns
+    -------
+    BytesIO
+        Bytes buffer of the ASDF file.
+    """
+    if asdffile_kwargs is None:
+        asdffile_kwargs = {}
+    if write_kwargs is None:
+        write_kwargs = {}
+
+    buff = BytesIO()
+    with asdf.AsdfFile(
+        tree, extensions=[WeldxExtension(), WeldxAsdfExtension()], **asdffile_kwargs
+    ) as ff:
+        ff.write_to(buff, **write_kwargs)
+        buff.seek(0)
+    return buff
+
+
+def _read_buffer(buffer: BytesIO, open_kwargs: dict = None):
+    """Read ASDF file contents from buffer instance.
+
+    Parameters
+    ----------
+    buffer
+        Buffer containing ASDF file contents
+    open_kwargs
+        Additional keywords to pass to asdf.AsdfFile.open()
+        Extensions are always set, copy_arrays=True is set by default.
+
+    Returns
+    -------
+    dict
+        ASDF file tree.
+
+    """
+    if open_kwargs is None:
+        open_kwargs = {"copy_arrays": True}
+
+    buffer.seek(0)
+    with asdf.open(
+        buffer, extensions=[WeldxExtension(), WeldxAsdfExtension()], **open_kwargs,
+    ) as af:
+        data = af.tree
+    return data
+
+
 def _write_read_buffer(
     tree: dict, asdffile_kwargs=None, write_kwargs=None, open_kwargs=None
 ):
     """Perform a buffered write/read roundtrip of a tree using default ASDF settings.
-
     Parameters
     ----------
     tree
@@ -291,34 +350,10 @@ def _write_read_buffer(
         Extensions are always set.
     open_kwargs
         Additional keywords to pass to asdf.AsdfFile.open()
-        Extensions and copy_arrays=True are always set.
-
+        Extensions are always set, copy_arrays=True is set by default.
     Returns
     -------
     dict
-
     """
-    # Write the data to buffer
-    if open_kwargs is None:
-        open_kwargs = {}
-    if write_kwargs is None:
-        write_kwargs = {}
-    if asdffile_kwargs is None:
-        asdffile_kwargs = {}
-
-    with asdf.AsdfFile(
-        tree, extensions=[WeldxExtension(), WeldxAsdfExtension()], **asdffile_kwargs
-    ) as ff:
-        buff = BytesIO()
-        ff.write_to(buff, **write_kwargs)
-        buff.seek(0)
-
-    # read back data from ASDF file contents in buffer
-    with asdf.open(
-        buff,
-        extensions=[WeldxExtension(), WeldxAsdfExtension()],
-        copy_arrays=True,
-        **open_kwargs,
-    ) as af:
-        data = af.tree
-    return data
+    buffer = _write_buffer(tree, asdffile_kwargs, write_kwargs)
+    return _read_buffer(buffer, open_kwargs)